Bug fixing

+++ tf_slim/pre_process_imagenet.py
@@ -270,8 +270,9 @@ class ImageCoder(object):
     return image
 
   def encode_jpeg(self, image_data):
+    image_data = image_data.astype(dtype=np.uint8)
     image = self._sess.run(self._encode_jpeg,
-                           feed_dict={self._encode_image_data: image_data.astype(dtype=np.uint8)})
+                           feed_dict={self._encode_image_data: image_data})
     return image
